Refactor: Move NoticeCard to components directory
Moved `frontend/src/NoticeCard.jsx` to `frontend/src/components/NoticeCard.jsx` to improve project organization and maintainability. Updated `frontend/src/components/NoticeFeed.jsx` to reflect the new import path.
