+++ frontend/src/components/NoticeCard.jsx
